+++ app/javascript/vue/tasks/otu/filter/App.vue
@@ -46,12 +46,25 @@
         <FilterList
           :list="list"
           :attributes="ATTRIBUTES"
+          :hide-unfrozen="hideFrozen"
           :preference-key="`tasks::filters::${OTU}`"
           v-model="selectedIds"
           radial-object
           @on-sort="list = $event"
           @remove="({ index }) => list.splice(index, 1)"
         />
+      </template>
+      <template #nav-settings-start>
+        <VToggle
+          title="Hide/show non-frozen columns"
+          @click="() => (hideFrozen = !hideFrozen)"
+        >
+          <VIcon
+            title="Hide/show non-frozen columns"
+            :name="hideFrozen ? 'contract' : 'expand'"
+            x-small
+          />
+        </VToggle>
       </template>
     </FilterLayout>
     <VSpinner
@@ -71,14 +84,18 @@ import useFilter from '@/shared/Filter/composition/useFilter.js'
 import RadialMatrix from '@/components/radials/matrix/radial.vue'
 import RadialOtu from '@/components/radials/otu/radial.vue'
 import VSpinner from '@/components/ui/VSpinner.vue'
+import VToggle from '@/components/ui/VToggle.vue'
+import VIcon from '@/components/ui/VIcon/index.vue'
 import { ATTRIBUTES } from './constants/attributes'
 import { listParser } from './utils/listParser'
 import { OTU } from '@/constants/index.js'
 import { Otu } from '@/routes/endpoints'
-import { computed } from 'vue'
+import { computed, ref } from 'vue'
 import csvDownload from './components/csvDownload.vue'
+import DwcChecklistDownload from './components/dwcChecklistDownload.vue'
 
 const extend = ['taxonomy']
+const hideFrozen = ref(false)
 
 const {
   append,
@@ -100,6 +117,15 @@ const extendDownload = computed(() => [
     component: csvDownload,
     bind: {
       params: parameters.value
+    }
+  },
+  {
+    label: 'DwC Checklist',
+    component: DwcChecklistDownload,
+    bind: {
+      params: parameters.value,
+      total: pagination.value?.total,
+      selectedIds: selectedIds.value
     }
   }
 ])