+++ config/routes/api_v1.rb
@@ -57,6 +57,7 @@ namespace :api, defaults: {format: :json} do
       get '/otus/inventory/alphabetical', to: '/otus#api_alphabetical_index', as: :api_alphabetical_index
       get '/otus/:id/inventory/content', to: '/otus#api_content', as: :api_content
       get '/otus/:id/inventory/distribution', to: '/otus#api_distribution', as: :api_distribution
+      get '/otus/:id/inventory/distribution_is_absent', to: '/otus#api_distribution_is_absent', as: :api_distribution_is_absent
       get '/otus/:id/inventory/keys', to: '/otus#api_key_inventory', as: :key_inventory
       get '/otus/:id/inventory/taxonomy', to: '/otus#api_taxonomy_inventory', as: :taxonomy_inventory
       get '/otus/:otu_id/inventory/images', to: '/otus#api_image_inventory', as: :otu_images_inventory