+++ config/routes/data.rb
@@ -34,6 +34,7 @@ resources :anatomical_parts do
     get :ontologies, defaults: {format: :json}
     get :children_of, defaults: {format: :json}
     get :ontology_autocomplete, defaults: {format: :json}
+    get :templates, defaults: {format: :json}
     get :used_ontologies, defaults: {format: :json}
 
     scope :select_ontologies, controller: 'tasks/anatomical_parts/select_ontologies' do
@@ -85,6 +86,7 @@ resources :biological_associations do
     get :autocomplete, defaults: {format: :json}
     get :select_options, defaults: {format: :json}
     get :subject_object_types, defaults: {format: :json}
+    get :origin_subject_index, defaults: {format: :json}
   end
 end
 
@@ -731,6 +733,7 @@ resources :otus do
     get :coordinate, defaults: {format: :json}
 
     get 'inventory/distribution', action: :distribution, defaults: {format: :json}
+    get 'inventory/distribution_is_absent', action: :distribution_is_absent, defaults: {format: :geojson}
     get 'inventory/taxonomy', action: :api_taxonomy_inventory, as: :taxonomy_inventory
     get 'inventory/citations', action: :citations_inventory, defaults: {format: :json}
   end