+++ CHANGELOG.md
@@ -9,12 +9,16 @@ This project <em>does not yet</em> adhere to [Semantic Versioning](https://semve
 
 ### Added
 
+- Distribution absence inference and mapping endpoints #4896
 - `subject_otu_id` and `object_otu_id` to biological associations `/api/v1/biological_associations/basic` endpoint
 - Filters: Copy to clipboard button for API Url [#4852]
+- Filters: Toggle hide/show non-frozen columns
 - In the New Asserted Distribution task, add map tab to Gazetteer selection for shape [#4853]
 - `/leads`, `/leads/:id/remaining_otus`and `/leads/:id/eliminated_otus` API endpoints
 - Filters: Save freeze columns state [#4882]
 - Filter Source: Add TaxonWorks CSL format to `download formatted` button [#4661]
+- DwCA names checklist export (from Filter OTUs as a download type) [#4643]
+- Create Anatomical Parts from the Biological Associations annotator slice [#4716]
 
 ### Changed
 
@@ -23,6 +27,10 @@ This project <em>does not yet</em> adhere to [Semantic Versioning](https://semve
 - OTU Smart Selector now sorts recent list by most recently used [#4865]
 - API endpoint for observation matrices only returns public ones instead of all
 - `Download formatted` modal now display italics in preview box [#1856]
+- Browse Taxon Name: Save `Redirect to valid name` state in user preferences [#4892]
+- New taxon name: Save `Autosave` state in user preferences
+- New source: Save `Autosave` state in user preferences
+- Remove all `user_id`s from api outputs [#4888]
 
 ### Fixed
 
@@ -40,9 +48,24 @@ This project <em>does not yet</em> adhere to [Semantic Versioning](https://semve
 - Browse OTU button creates OTUs with an incorrect taxon name ID on the /otus/:id page [#4879]
 - The Browse taxon name task bar is positioned incorrectly when the header bar is fixed.
 - Filter Source: The style selection in formatted download always applies the same style regardless of what is selected.
+- Fix some tests that intermittently fail in our CI (Continuous Integration) environment, add a script to produce a report of such failures [#4891]
+- Added `otu_id` to DwcOccurrence model [#4767]
+- Improved speed of OTU distribution responses
+- [#2141] On the Browse OTU page, display Biological Associations related to an OTU by:
+  - Biological Association -> subject/object is a Collection Object or Field Occurrence -> the Taxon Determination of the Collection Object or Field Occurrence is the OTU in question
+  - Anatomical Part has origin OTU the OTU in question
+- ISBNs with 7-digit parts not recognized as valid [#4819]
+- Batch Source Import returns 'Serial identifiers identifier has already been taken' with no clear reason or explanation
+- Pin source button is missing in New Image task
 
+[#4896]: https://github.com/SpeciesFileGroup/taxonworks/issues/4896
 [#1856]: https://github.com/SpeciesFileGroup/taxonworks/issues/1856
+[#2141]: https://github.com/SpeciesFileGroup/taxonworks/issues/2141
+[#4643]: https://github.com/SpeciesFileGroup/taxonworks/issues/4643
 [#4661]: https://github.com/SpeciesFileGroup/taxonworks/issues/4661
+[#4716]: https://github.com/SpeciesFileGroup/taxonworks/issues/4716
+[#4767]: https://github.com/SpeciesFileGroup/taxonworks/issues/4767
+[#4819]: https://github.com/SpeciesFileGroup/taxonpages/issues/4819
 [#4851]: https://github.com/SpeciesFileGroup/taxonpages/issues/4851
 [#4852]: https://github.com/SpeciesFileGroup/taxonpages/issues/4852
 [#4853]: https://github.com/SpeciesFileGroup/taxonpages/issues/4853
@@ -51,6 +74,9 @@ This project <em>does not yet</em> adhere to [Semantic Versioning](https://semve
 [#4866]: https://github.com/SpeciesFileGroup/taxonpages/issues/4866
 [#4879]: https://github.com/SpeciesFileGroup/taxonpages/issues/4879
 [#4882]: https://github.com/SpeciesFileGroup/taxonpages/issues/4882
+[#4888]: https://github.com/SpeciesFileGroup/taxonpages/issues/4888
+[#4892]: https://github.com/SpeciesFileGroup/taxonpages/issues/4892
+[#4891]: https://github.com/SpeciesFileGroup/taxonpages/issues/4891
 
 ## [0.61.0] - 2026-02-20
 
@@ -108,6 +134,7 @@ This project <em>does not yet</em> adhere to [Semantic Versioning](https://semve
 - Filters: keep per state after reset [#4831]
 - Updated Ruby gems
 - Refactor Filter annotations task [#4809]
+- Focus the next keyboard input after mouse clicks in New Taxon Name
 
 [#345]: https://github.com/SpeciesFileGroup/taxonpages/issues/345
 [#4271]: https://github.com/SpeciesFileGroup/taxonworks/issues/4271